Added pagination to Recipes and started setting up Access control

+++ Project/ExpirationManager/src/main/java/softuni/expirationManager/web/CategoryController.java
@@ -2,12 +2,15 @@ package softuni.expirationManager.web;
 
 import jakarta.validation.Valid;
 import org.apache.tomcat.util.http.fileupload.impl.FileSizeLimitExceededException;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import softuni.expirationManager.model.MyUserDetails;
 import softuni.expirationManager.model.dtos.category.CategoryAddDTO;
 import softuni.expirationManager.model.dtos.category.CategoryEditDTO;
 import softuni.expirationManager.model.dtos.category.CategoryViewDTO;
@@ -64,7 +67,11 @@ public class CategoryController {
     }
 
     @GetMapping("/categories/{id}/edit")
-    public String getEditCategory(@PathVariable Long id, Model model) {
+    public String getEditCategory(@PathVariable Long id, Model model, @AuthenticationPrincipal MyUserDetails userDetails) {
+
+        if (isNotAuthorized(userDetails) && this.categoryService.isNotOwner(userDetails.getId(), id)) {
+            throw new AccessDeniedException("Access denied");
+        }
 
         if (!model.containsAttribute("categoryEditDTO")) {
             model.addAttribute("categoryEditDTO", this.categoryService.getCategoryEditDtoById(id));
@@ -76,7 +83,12 @@ public class CategoryController {
     @PutMapping("/categories/{id}/edit")
     public String putEditCategory(@Valid CategoryEditDTO categoryEditDTO,
                                   BindingResult bindingResult,
-                                  RedirectAttributes redirectAttributes) throws IOException {
+                                  RedirectAttributes redirectAttributes,
+                                  @AuthenticationPrincipal MyUserDetails userDetails) throws IOException {
+
+        if (isNotAuthorized(userDetails)) {
+            throw new AccessDeniedException("Access denied");
+        }
 
         if (bindingResult.hasErrors()) {
             redirectAttributes.addFlashAttribute("categoryEditDTO", categoryEditDTO);
@@ -91,14 +103,24 @@ public class CategoryController {
     }
 
     @DeleteMapping("/categories/{id}")
-    public String deleteCategory(@PathVariable Long id) {
+    public String deleteCategory(@PathVariable Long id, @AuthenticationPrincipal MyUserDetails userDetails) {
+
+        if (isNotAuthorized(userDetails) && this.categoryService.isNotOwner(userDetails.getId(), id)) {
+            throw new AccessDeniedException("Access denied");
+        }
+
         this.categoryService.deleteById(id);
 
         return "redirect:/categories";
     }
 
     @GetMapping("/categories/{id}")
-    public String getProductsByCategory(@PathVariable Long id, Model model) {
+    public String getProductsByCategory(@PathVariable Long id, Model model, @AuthenticationPrincipal MyUserDetails userDetails) {
+
+        if (isNotAuthorized(userDetails) && this.categoryService.isNotOwner(userDetails.getId(), id)) {
+            throw new AccessDeniedException("Access denied");
+        }
+
         model.addAttribute("category", this.categoryService.getCategoryNameIdDTO(id));
 
         return "category";
@@ -112,5 +134,9 @@ public class CategoryController {
         mav.addObject("errorMessage", "File you uploaded exceeds maximum allowed size");
 
         return mav;
+    }
+
+    private boolean isNotAuthorized(MyUserDetails userDetails) {
+        return userDetails.getAuthorities().stream().noneMatch(a -> "ROLE_ADMIN".equals(a.getAuthority()));
     }
 }
